third_party/blink/web_tests/fast/canvas-api/canvas-getImageData-large-crash.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(t) {
    var canvas = document.createElement("canvas");
    assert_throws_js(TypeError, function() {
        canvas.getContext("2d").getImageData(10, 0xffffffff, 2147483647, 10); });
}, 'Test that canvas crash when large image data cannot be allocated.');
</script>
